requirements.txt
dill >= 0.3.7
hilbertcurve >= 2.0.5
matplotlib >= 3.3.4
numpy >= 1.21.6
opencv-python >= 4.9.0.80
pandas >= 1.3.5
pefile >= 2023.2.7
Pillow >= 9.5.0
PyWavelets >= 1.3.0
scikit-learn >= 0.24.1
torch >= 1.13.1
torchvision >= 0.14.1
